reorganize tests files of create authorization
(cherry picked from commit b557396b744e383bddefecfb9411ee51e9ec4b2b)
