+++ packages/graphql/tests/integration/mutations/create/authorization/validate/create-authorization-validate-union.test.ts
@@ -18,8 +18,8 @@
  */
 
 import { GraphQLError } from "graphql";
-import type { UniqueType } from "../../../../utils/graphql-types";
-import { TestHelper } from "../../../../utils/tests-helper";
+import type { UniqueType } from "../../../../../utils/graphql-types";
+import { TestHelper } from "../../../../../utils/tests-helper";
 
 describe("create interface with authorization validate", () => {
     const testHelper = new TestHelper();
@@ -34,17 +34,15 @@ describe("create interface with authorization validate", () => {
         Comment = testHelper.createUniqueType("Comment");
 
         const typeDefs = /* GraphQL */ `
-            interface Content {
-                title: String!
-            }
+            union Content = ${Post} | ${Comment}
 
-            type ${Post} implements Content @node {
+            type ${Post} @node {
                 title: String!
                 creatorId: ID!
                 creator: [${User}!]! @relationship(type: "HAS_CONTENT", direction: IN, properties: "CreatedAt")
             }
 
-            type ${Comment} implements Content @node {
+            type ${Comment} @node {
                 title: String!
                 episodes: Int
                 creator: [${User}!]! @relationship(type: "HAS_CONTENT", direction: IN, properties: "CreatedAt")
@@ -94,7 +92,7 @@ describe("create interface with authorization validate", () => {
                         {
                             id: $id
                             name: "Marvin"
-                            content: { create: [{ node: { ${Post}: { title: "Doors are evil", creatorId: $id } } }] }
+                            content: { ${Post}: {create: { node: { title: "Doors are evil", creatorId: $id } } } }
                         }
                     ]
                 ) {
@@ -146,7 +144,7 @@ describe("create interface with authorization validate", () => {
                         {
                             id: $id
                             name: "Marvin"
-                            content: { create: [{ node: { ${Post}: { title: "Doors are evil", creatorId: "another-id" } } }] }
+                            content: { ${Post}: { create: [{ node: { title: "Doors are evil", creatorId: "another-id" }  }]} }
                         }
                     ]
                 ) {
@@ -177,7 +175,7 @@ describe("create interface with authorization validate", () => {
                         {
                             id: $id
                             name: "Marvin"
-                            content: { connect: { where: { node: { title: { eq: "Doors are evil" } } }} }
+                            content: { ${Post}: { connect: { where: { node: { title: { eq: "Doors are evil" } } }} } }
                         }
                     ]
                 ) {
@@ -231,7 +229,7 @@ describe("create interface with authorization validate", () => {
                         {
                             id: $id
                             name: "Marvin"
-                            content: { connect: { where: { node: { title: { eq: "Doors are evil" } } }} }
+                            content: { ${Post}: { connect: { where: { node: { title: { eq: "Doors are evil" } } }} } }
                         }
                     ]
                 ) {
@@ -264,7 +262,7 @@ describe("create interface with authorization validate", () => {
                         {
                             id: $id
                             name: "Marvin"
-                            content: { connect: { where: { node: { title: { eq: "Doors are evil" } } }} }
+                            content: { ${Comment}: { connect: { where: { node: { title: { eq: "Doors are evil" } } }} } }
                         }
                     ]
                 ) {
