+++ packages/graphql/tests/integration/mutations/create/authorization/validate/create-authorization-validate.int.test.ts
@@ -18,8 +18,8 @@
  */
 
 import { GraphQLError } from "graphql";
-import type { UniqueType } from "../../../../utils/graphql-types";
-import { TestHelper } from "../../../../utils/tests-helper";
+import type { UniqueType } from "../../../../../utils/graphql-types";
+import { TestHelper } from "../../../../../utils/tests-helper";
 
 describe("create with authorization validate", () => {
     const testHelper = new TestHelper();
